REFACTOR: addressed review comments
